Cost column added in Stouk out report.

+++ Pspcl/Pspcl.Services/Interfaces/IStockService.cs
@@ -32,7 +32,7 @@ namespace Pspcl.Services.Interfaces
 
         //public int GetCorrespondingRateValue(int materialId);
         public bool isGrnNumberExist(string GrnNumber);
-        public int GetCost(int materialId,int noOfUnits);
+        public float GetCost(int materialId,int noOfUnits);
 
         public bool srNoValidationInDatabase(List<int> serialNumbers, int materialGroupId,int materialTypeId, int materialId, string make);
 		public Dictionary<String, int> AllMakesAndQuantities(int materialGroupId, int materialTypeId, int materialId);
